# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: YAFA VANAM Commerce API
  version: 2.0.0
  description: |
    The full HTTP contract of the Go commerce API: catalogue, carts, orders,
    Razorpay payments and session auth.

    Conventions:

    - Commerce resources live under `/api/v1`; session auth lives at the root
      (`/auth/...`) because the storefront proxies it separately.
    - Errors use two shapes. Commerce handlers return
      `{"error":{"code","message"}}`; auth handlers return `{"error":"message"}`.
    - Browser mutations while signed in must send `X-CSRF-Token` (double-submit
      with the `yafa_csrf` cookie obtained from `GET /auth/csrf`).
    - Money is a decimal number in INR (e.g. 2119.0); Razorpay amounts are paise.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/products:
    get:
      tags: [Catalogue]
      summary: Search and filter active products
      parameters:
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: subcategory, schema: { type: string } }
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
        - { in: query, name: offset, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductList" }
  /api/v1/products/{slug}:
    get:
      tags: [Catalogue]
      summary: Get an active product with its variants
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/carts:
    post:
      tags: [Carts]
      summary: Create a cart
      description: Anonymous by default; bound to the signed-in account when a session cookie is present. A signed-in browser must send X-CSRF-Token.
      security:
        - cookieAuth: []
        - {}
      responses:
        "201":
          description: New cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
  /api/v1/carts/{cartId}:
    get:
      tags: [Carts]
      summary: Read a cart with server-calculated prices
      description: Guests may read their own cart by ID; a signed-in user may read carts bound to their account. A claimed cart is no longer guest-accessible.
      security:
        - cookieAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/CartId"
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/carts/{cartId}/items:
    post:
      tags: [Carts]
      summary: Add a validated product variant to a cart
      description: Prices and stock come from the server-side catalogue; quantity is capped at 20 per line. If the caller is signed in, an anonymous cart is claimed for their account on first mutation.
      security:
        - cookieAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/CartId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [product_id, variant_id, quantity]
              properties:
                product_id: { type: string }
                variant_id: { type: string }
                quantity: { type: integer, minimum: 1, maximum: 20 }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VariantUnavailable" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/v1/carts/{cartId}/items/{variantId}:
    parameters:
      - $ref: "#/components/parameters/CartId"
      - in: path
        name: variantId
        required: true
        schema: { type: string }
    patch:
      tags: [Carts]
      summary: Set a cart line quantity (1-20)
      security:
        - cookieAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1, maximum: 20 }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VariantUnavailable" }
        "422": { $ref: "#/components/responses/Validation" }
    delete:
      tags: [Carts]
      summary: Remove a cart line
      security:
        - cookieAuth: []
        - {}
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List the signed-in user's orders (newest first)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Order list; access tokens are never included
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/SimpleUnauthorized" }
    post:
      tags: [Orders]
      summary: Create a pending-payment order from a server-priced cart
      description: |
        Requires sign-in in the production configuration. Shipping is 199
        standard, 299 express, or free when the discounted subtotal is at least
        1999. Retries must reuse the same Idempotency-Key: the first call
        returns 201, a replay returns the original order with 200 and an
        `Idempotent-Replayed: true` header.
      security:
        - cookieAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrder" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "200":
          description: Existing order returned after an idempotent replay
          headers:
            Idempotent-Replayed:
              schema: { type: string, const: "true" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/SimpleUnauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/v1/orders/{orderNumber}:
    get:
      tags: [Orders]
      summary: Read one order
      description: |
        The signed-in owner reads their order with the session cookie. When the
        API runs without auth configured (local development only), guests read
        an order with the X-Order-Access-Token header returned at creation.
      security:
        - cookieAuth: []
        - orderToken: []
      parameters:
        - in: path
          name: orderNumber
          required: true
          schema: { type: string, pattern: "^YV-[0-9]{8}-[A-Z0-9]{4}$" }
        - in: header
          name: X-Order-Access-Token
          schema: { type: string }
      responses:
        "200":
          description: Order (access token never echoed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "403": { $ref: "#/components/responses/AccessDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/payments/razorpay/orders:
    post:
      tags: [Payments]
      summary: Create (or reuse) a Razorpay order for a checkout
      description: Creates the local pending-payment order, then registers it with Razorpay and attaches the gateway ID. Requires Razorpay configuration; otherwise 503.
      security:
        - cookieAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          schema: { type: string, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrder" }
      responses:
        "201":
          description: Razorpay checkout parameters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RazorpayCheckoutOrder" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/SimpleUnauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
        "502":
          description: Razorpay gateway call failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: Payments not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/payments/razorpay/verify:
    post:
      tags: [Payments]
      summary: Verify a checkout's HMAC signature and mark the payment authorized
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RazorpayVerificationInput" }
      responses:
        "200":
          description: Signature valid; payment recorded as AUTHORIZED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RazorpayVerificationResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/SimpleUnauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: yafa_access
      description: HTTP-only JWT session cookie issued by /auth/login, /auth/register, /auth/refresh, or the Google callback.
    orderToken:
      type: apiKey
      in: header
      name: X-Order-Access-Token
      description: Guest order credential returned as access_token when the order is created (no-auth mode only).
  parameters:
    CartId:
      in: path
      name: cartId
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AccessDenied:
      description: The resource belongs to someone else or a credential is missing/invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Product:
      type: object
      required:
        [id, name, slug, brand, category, subcategory, product_type, status, description, commerce, variants, images, benefits, usage, warnings, ingredients]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        brand: { type: string }
        category: { type: string }
        subcategory: { type: string }
        product_type: { type: string }
        status: { type: string, enum: [active, draft, archived] }
        description: { $ref: "#/components/schemas/Description" }
        commerce: { $ref: "#/components/schemas/CommerceMoney" }
        variants:
          type: array
          items: { $ref: "#/components/schemas/Variant" }
        images: { $ref: "#/components/schemas/Images" }
        benefits: { type: array, items: { type: string } }
        usage: { $ref: "#/components/schemas/Usage" }
        warnings: { type: array, items: { type: string } }
        ingredients: { $ref: "#/components/schemas/Ingredients" }
    ProductList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        total: { type: integer, minimum: 0 }
        limit: { type: integer }
        offset: { type: integer, minimum: 0 }
...
    Cart:
      type: object
      required: [id, items, item_count, subtotal, currency, updated_at]
      properties:
        id: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/CartLine" }
        item_count: { type: integer, minimum: 0 }
        subtotal: { type: number, minimum: 0 }
        currency: { type: string }
        updated_at: { type: string, format: date-time }
...